chore: create the invite model and tweak the other models' relationships
