+++ backend/feedback/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import *
+
+# Register your models here.
+admin.site.register(KataFeedback)
+admin.site.register(SFeedback)
+admin.site.register(MCFeedback)
